Refactor json serializers
And move mapper tests to core
