+++ Letterbook.Api.IntegrationTests/ProfileTests.cs
@@ -4,12 +4,12 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using AutoMapper;
 using Letterbook.Api.IntegrationTests.Fixtures;
-using Letterbook.Api.Json;
 using Letterbook.Api.Tests.Fakes;
 using Letterbook.Core.Extensions;
 using Letterbook.Core.Models;
 using Letterbook.Core.Models.Dto;
 using Letterbook.Core.Models.Mappers;
+using Letterbook.Core.Models.Mappers.Converters;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using Profile = Letterbook.Core.Models.Profile;
