+++ Letterbook.Core.Tests/MapperTests.cs
@@ -1,12 +1,10 @@
 using AutoMapper;
-using Letterbook.Api.Dto;
 using Letterbook.Core.Models.Dto;
 using Letterbook.Core.Models.Mappers;
-using Letterbook.Core.Tests;
 using Letterbook.Core.Tests.Fakes;
 using Medo;
 
-namespace Letterbook.Api.Tests;
+namespace Letterbook.Core.Tests;
 
 public class MapperTests : WithMocks
 {
